Переписала Most Commented и Top Rated на компонент
