+++ src/components/film-list-extra.js
@@ -1,27 +1,10 @@
 import {createFilmTemplate} from './film.js';
 import {createElement} from '../util.js';
 
-const EXTRA_TITLES = [`Top Rated`, `Most Commented`];
+const createFilmListExtraTemplate = (title, sortedFilms) => {
 
-const getMostCommentedFilms = (films) => {
-  return films.slice().sort((film1, film2) => {
-    return film2.comments.length - film1.comments.length;
-  })
-.slice(0, 2);
-};
-
-const getMostRatedFilms = (films) => {
-  return films.slice()
- .sort((film1, film2) => {
-   return film2.rating - film1.rating;
- })
-.slice(0, 2);
-};
-
-const createFilmListExtraMarkup = (extraFilm) => {
-  const {title, topFilms} = extraFilm;
-  const sortedFilmsMarkup = topFilms.map((it) => {
-    return createFilmTemplate(it);
+  const sortedFilmsMarkup = sortedFilms.map((sortedFilm) => {
+    return createFilmTemplate(sortedFilm);
   }).join(`\n`);
 
   return (
@@ -34,39 +17,15 @@ const createFilmListExtraMarkup = (extraFilm) => {
   );
 };
 
-
-const createFilmListExtraTemplate = (films) => {
-
-  const topFilmsMap = {
-    'Top Rated': getMostRatedFilms(films),
-    'Most Commented': getMostCommentedFilms(films)
-  };
-
-
-  const generateExtraFilms = () => {
-    return EXTRA_TITLES.map((title) => {
-      return {
-        title,
-        topFilms: topFilmsMap[title]
-      };
-    });
-  };
-
-  const extraFilms = generateExtraFilms();
-
-  return extraFilms.map((it) => {
-    return createFilmListExtraMarkup(it);
-  }).join(`\n`);
-};
-
 export default class FilmsListExtra {
-  constructor(films) {
+  constructor(title, films) {
+    this._title = title;
     this._films = films;
     this._element = null;
   }
 
   getTemplate() {
-    return createFilmListExtraTemplate(this._films);
+    return createFilmListExtraTemplate(this._title, this._films);
   }
 
   getElement() {
